transport: promote BBox to shared as a pure spatial type
Move BBox out of transport::groups into shared::bbox so the upcoming
motis crate can reuse it. transport imports shared::BBox directly.
